requirements.txt
# LangChain 和 LLM 相关
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# Web 爬虫
playwright>=1.40.0
nest-asyncio>=1.5.0

# API 服务
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
websockets>=11.0

# 数据库和认证
sqlalchemy>=2.0.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pydantic[email]>=2.0.0

# 工具库
python-dateutil>=2.8.2

# 阿里云短信服务
alibabacloud-dysmsapi20170525>=2.0.24
alibabacloud-tea-openapi>=0.3.7
